Move root level files into src
